separate Activity.vue into view & component
