+++ frontend/src/locales/de.json
@@ -15,6 +15,12 @@
       },
       "menuCardlessContentNode": {
         "deletingDisabled": "Muss leer sein zum Löschen"
+      },
+      "scheduleEntry": {
+        "backToContents": "Zurück zum Bearbeiten des Inhalts",
+        "changeLayout": "Layout ändern",
+        "deleteWarning": "Möchtest du diese Aktivität wirklich löschen? Der komplette Inhalt dieser Aktivität wird gelöscht.",
+        "layout": "Layout"
       }
     },
     "camp": {
@@ -286,12 +292,6 @@
   },
   "views": {
     "activity": {
-      "activity": {
-        "backToContents": "Zurück zum Bearbeiten des Inhalts",
-        "changeLayout": "Layout ändern",
-        "deleteWarning": "Möchtest du diese Aktivität wirklich löschen? Der komplette Inhalt dieser Aktivität wird gelöscht.",
-        "layout": "Layout"
-      },
       "category": {
         "backToContents": "Zurück zum Bearbeiten des Inhalts",
         "changeLayout": "Layout ändern",
