+++ frontend/src/locales/en.json
@@ -15,6 +15,12 @@
       },
       "menuCardlessContentNode": {
         "deletingDisabled": "Must be empty to delete"
+      },
+      "scheduleEntry": {
+        "backToContents": "Back to editing content",
+        "changeLayout": "Change layout",
+        "deleteWarning": "Do you really want to delete this activity? All content of this activity will be removed.",
+        "layout": "Layout"
       }
     },
     "camp": {
@@ -283,12 +289,6 @@
   },
   "views": {
     "activity": {
-      "activity": {
-        "backToContents": "Back to editing content",
-        "changeLayout": "Change layout",
-        "deleteWarning": "Do you really want to delete this activity? All content of this activity will be removed.",
-        "layout": "Layout"
-      },
       "category": {
         "backToContents": "Back to editing content",
         "changeLayout": "Change layout",
